Add Startup Setting to make Red/Green Wires Optional
The addition of Green/Red wires to Circuit Network items made them more troublesome to obtain, as well as proved conflicting with the Wire Shortcuts mod. CN items are dificult in their application, not obtaining, thus the additional complexity seemed counter-productive to encouraging the use of CN in complex setups.